+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Release 5.7.6 (unreleased)
+ * Credentials: Fix displaying age verification credential
 
 # Release 5.7.5
  * Update to VC-K 5.10.1